+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning].
     - Add JSON serialization support for `Error` class.
     - Add `ExpandableError` class, designed to losslessly capture additional properties when deserialized.
     - Add `WithIdentifier` extension methods for `Error` and `ExpandableError`.
+    - Add `OnFail` extension methods for all result types and `OnNone` extension methods for `Maybe<T>`.
 
 ## [1.0.0-alpha09] - 2022-05-19